+++ src/coveralls.ts
@@ -1,6 +1,6 @@
 import * as core from "@actions/core";
 import * as fs from "fs";
-import * as http from "./http.mjs";
+import * as http from "./http.js";
 
 export async function patch(coverallsOut: string) {
   let data: string = fs.readFileSync(coverallsOut).toString();
